Refactoring of app reducers
