+++ src/reducer/name-spaces.js
@@ -1,0 +1,7 @@
+
+const NameSpaces = {
+  Catalog: `catalog`,
+};
+
+export default NameSpaces;
+
